rpk: move cmd/topic to its own package
